[Sound]: Noticed some 32 bit frame counting that is now properly done 64 bit wise. [CLGame]: Added 'client_env_sound' entity similar concept to what Half-Life has.

+++ inc/client/sound/dma.h
@@ -51,7 +51,7 @@ typedef struct {
 } snddma_driver_t;
 
 extern dma_t    dma;
-extern int      s_paintedtime;
+extern int64_t  s_paintedtime;
 
 extern cvar_t   *s_khz;
